[Refactor] Extract jax runner tests into their own test file.
